Task-6/requirements.txt
streamlit==1.33.0
google-generativeai==0.3.2
Pillow==10.2.0
python-dotenv==1.0.1
PyMuPDF==1.23.25
